LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-margin-focus-expected.txt

This header may obscure content below it.
Focus this button
Start on this button

FAIL Focus should scroll button outside of scroll margin assert_equals: scrolls to bottom of scroller expected 69 but got 0
